requirements.txt
Django==4.2.16
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
replicate==0.26.0
